requirements.txt
einops==0.8.0
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.8.1.78
PyYAML==6.0.2
scipy==1.14.1
timm==0.9.12
torch==2.1.2
tqdm==4.66.1
